chore: begin refactoring out testids
