+++ frontend/routes/onboarding/create-wallet/create-wallet.spec.tsx
@@ -1,7 +1,7 @@
 import { MemoryRouter } from 'react-router-dom'
-import { CreateWallet } from '.'
-import locators from '../../../components/locators'
-import { createNewWalletButton, importWalletButton } from '../../../locator-ids'
+import { CreateWallet, locators } from '.'
+import componentLocators from '../../../components/locators'
+
 import { fireEvent, render, screen } from '@testing-library/react'
 import { FULL_ROUTES } from '../../route-names'
 
@@ -21,21 +21,21 @@ const renderComponent = () =>
 describe('Create wallet', () => {
   it('renders icon and buttons', () => {
     renderComponent()
-    expect(screen.getByTestId(locators.walletIcon)).toBeInTheDocument()
-    expect(screen.getByTestId(createNewWalletButton)).toHaveTextContent('Create a wallet')
-    expect(screen.getByTestId(createNewWalletButton)).toHaveFocus()
-    expect(screen.getByTestId(importWalletButton)).toHaveTextContent('Import a Wallet')
+    expect(screen.getByTestId(componentLocators.walletIcon)).toBeInTheDocument()
+    expect(screen.getByTestId(locators.createNewWalletButton)).toHaveTextContent('Create a wallet')
+    expect(screen.getByTestId(locators.createNewWalletButton)).toHaveFocus()
+    expect(screen.getByTestId(locators.importWalletButton)).toHaveTextContent('Import a Wallet')
   })
 
   it('navigates to import wallet page when import button is pressed', () => {
     renderComponent()
-    fireEvent.click(screen.getByTestId(importWalletButton))
+    fireEvent.click(screen.getByTestId(locators.importWalletButton))
     expect(mockNavigate).toBeCalledWith(FULL_ROUTES.importWallet)
   })
 
   it('navigates to save mnemonic route when create wallet button is pressed', async () => {
     renderComponent()
-    fireEvent.click(screen.getByTestId(createNewWalletButton))
+    fireEvent.click(screen.getByTestId(locators.createNewWalletButton))
     expect(mockNavigate).toBeCalledWith(FULL_ROUTES.saveMnemonic)
   })
 })
