+++ test/e2e/page-objects/get-started.ts
@@ -1,6 +1,6 @@
 import { By, WebDriver } from 'selenium-webdriver'
 import { clickElement, getByDataTestID, isElementDisplayed } from '../helpers/selenium-util'
-import * as locators from '../../../frontend/locator-ids'
+import { locators } from '../../../frontend/routes/onboarding/get-started'
 import { Password } from './password'
 
 export class GetStarted {
